Fix unhelpful errors in untitled workspaces
This change fixes an issue where any command that attempted to infer
the workspace manifest would fail with vague errors when in a untitled
workspace.

This is also the beginning of a refactor of error handling in internal
functions.  Currently when a internal function throws an error it uses
PSCmdlet.ThrowTerminatingError through the internal function
ThrowError. This makes it difficult for ThrowError to reliably determine
which context to throw from, and makes it impossible to catch. Internal
functions will now use the throw statement with an error record so the
caller can rethrow or suppress.