Better error handling in untitled workspaces and internal functions (#22)
* Fix unhelpful errors in untitled workspaces

This change fixes an issue where any command that attempted to infer
the workspace manifest would fail with vague errors when in a untitled
workspace.

This is also the beginning of a refactor of error handling in internal
functions.  Currently when a internal function throws an error it uses
PSCmdlet.ThrowTerminatingError through the internal function
ThrowError. This makes it difficult for ThrowError to reliably determine
which context to throw from, and makes it impossible to catch. Internal
functions will now use the throw statement with an error record so the
caller can rethrow or suppress.

* Switch GetInferredMember function

Also remove the checking for arguments in the AST and the logic
that used it in Type.FindMembers.

* Switch GetInferredType and GetType

* Switch SetEditorLocation

* Fix error preference gathering in ThrowError

Move the preference check to after context is retrieved so we can
check the preference of the caller.

* More fixes for untitled workspaces

- Fix additional parameter binding issues in GetSettings
- Remove CmdletBinding from GetancestorOrThrow.  This internal function
  will still call ThrowError and CmdletBinding will cause ThrowError to
  throw in the wrong context

* Fix workspace path resolution and resources

Some of the changes broke resolving paths relative to the workspace
path, this change fixes that.  Also add some additional handling
to ConvertTo-LocalizationString that accounts for the file not
existing yet.